Finally updated the log.
I've only been updating my log on twitter for the last weeks so I've only copy/pasted this. But from now on I'd like to write a more detailed log with examples of codes, so I can refer to it later if I want to study things we've learned at school again in the future.